Consolidate more tests
Consolidate more tests from their own dedicated test binaries into
existing test suites. Note that our backtrace tests need a bit of a
special treatment here, because they start flaking with this change. The
reason seems to be that the standard library employs caching of the
environment variable contents (without any API for invalidation...)
influencing backtrace capture and now unrelated tests could pre-populate
this cache by just attempting to capture a backtrace. Hence, we have to
resort to reimplementing the environment variable evaluation logic for
testing purposes. Ouch.
