+++ src/error.rs
@@ -2,6 +2,7 @@ use std::backtrace::Backtrace;
 use std::backtrace::BacktraceStatus;
 use std::borrow::Borrow;
 use std::borrow::Cow;
+use std::env;
 use std::error::Error as StdError;
 use std::fmt::Debug;
 use std::fmt::Display;
@@ -95,6 +96,39 @@ impl IntoCowStr for String {
         Cow::Owned(self.into_boxed_str())
     }
 }
+
+
+trait MaybeCapture {
+    fn maybe_capture() -> Backtrace;
+}
+
+impl MaybeCapture for Backtrace {
+    fn maybe_capture() -> Backtrace {
+        if cfg!(debug_assertions) {
+            // For testing purposes we need to circumvent the standard
+            // library's caching logic, which otherwise could cause
+            // flakiness in our tests, as they'd be impacted by
+            // unrelated errors that check for backtrace capture.
+            let enabled = match env::var("RUST_LIB_BACKTRACE") {
+                Ok(s) => s != "0",
+                Err(_) => match env::var("RUST_BACKTRACE") {
+                    Ok(s) => s != "0",
+                    Err(_) => false,
+                },
+            };
+            if enabled {
+                Backtrace::force_capture()
+            } else {
+                Backtrace::disabled()
+            }
+        } else {
+            // Just use the standard capture infrastructure and its caching
+            // logic.
+            Self::capture()
+        }
+    }
+}
+
 
 enum ErrorImpl {
     #[cfg(feature = "dwarf")]
@@ -529,7 +563,7 @@ impl From<gimli::Error> for Error {
             error: Box::new(ErrorImpl::Dwarf {
                 error: other,
                 #[cfg(feature = "backtrace")]
-                backtrace: Backtrace::capture(),
+                backtrace: Backtrace::maybe_capture(),
             }),
         }
     }
@@ -541,7 +575,7 @@ impl From<io::Error> for Error {
             error: Box::new(ErrorImpl::Io {
                 error: other,
                 #[cfg(feature = "backtrace")]
-                backtrace: Backtrace::capture(),
+                backtrace: Backtrace::maybe_capture(),
             }),
         }
     }
@@ -557,7 +591,7 @@ impl From<Box<dyn StdError + Send + Sync + 'static>> for Error {
                 // because no trait exposes it. So the best we can do is capture
                 // a backtrace ourselves. Sigh.
                 #[cfg(feature = "backtrace")]
-                backtrace: Backtrace::capture(),
+                backtrace: Backtrace::maybe_capture(),
             }),
         }
     }
@@ -725,8 +759,10 @@ impl<T> IntoError<T> for Option<T> {
 mod tests {
     use super::*;
 
+    use std::env;
     use std::mem::size_of;
 
+    use test_log::test as forked_test;
     use test_log::test;
     use test_tag::tag;
 
@@ -854,5 +890,58 @@ Caused by:
     some invalid data"#;
         assert_eq!(format!("{err:?}"), expected);
         assert_ne!(format!("{err:#?}"), "");
+    }
+
+    /// Make sure that we can capture backtraces in errors.
+    ///
+    /// # Notes
+    /// This test requires sufficient debug information to be present so
+    /// that the file name is contained in the backtrace. For that reason we
+    /// only run it on debug builds (represented by the `debug_assertions`
+    /// proxy cfg).
+    #[forked_test]
+    fn error_backtrace() {
+        if !cfg!(debug_assertions) {
+            return
+        }
+
+        // Ensure that we capture a backtrace.
+        let () = env::set_var("RUST_LIB_BACKTRACE", "1");
+
+        let err = io::Error::new(io::ErrorKind::InvalidData, "some invalid data");
+        let err = Error::from(err);
+        let debug = format!("{err:?}");
+
+        let start_idx = debug.find("Stack backtrace").unwrap();
+        let backtrace = &debug[start_idx..];
+        assert!(backtrace.contains("src/error.rs"), "{backtrace}");
+    }
+
+    /// Make sure that we do not emit backtraces in errors when
+    /// the `RUST_LIB_BACKTRACE` environment variable is not present.
+    #[forked_test]
+    fn error_no_backtrace1() {
+        let () = env::remove_var("RUST_BACKTRACE");
+        let () = env::remove_var("RUST_LIB_BACKTRACE");
+
+        let err = io::Error::new(io::ErrorKind::InvalidData, "some invalid data");
+        let err = Error::from(err);
+        let debug = format!("{err:?}");
+
+        assert_eq!(debug.find("Stack backtrace"), None);
+    }
+
+    /// Make sure that we do not emit backtraces in errors when
+    /// the `RUST_LIB_BACKTRACE` environment variable is "0".
+    #[forked_test]
+    fn error_no_backtrace2() {
+        let () = env::remove_var("RUST_BACKTRACE");
+        let () = env::set_var("RUST_LIB_BACKTRACE", "0");
+
+        let err = io::Error::new(io::ErrorKind::InvalidData, "some invalid data");
+        let err = Error::from(err);
+        let debug = format!("{err:?}");
+
+        assert_eq!(debug.find("Stack backtrace"), None);
     }
 }
